feat: add withdrawal reject and view dialogs, update withdrawal table and schema
- Implemented `WithdrawalRejectDialog` for rejecting withdrawal requests with an admin note.
- Added `WithdrawalViewDialog` to display detailed information about withdrawal requests.
- Updated `WithdrawalTable` to improve column visibility and search functionality.
- Modified withdrawal request schema to include additional user information (fullName, email, avatarUrl).
- Changed API mutation methods for approving and rejecting withdrawals to use PATCH instead of POST.
- Removed unused routes related to companies, tasks, and templates from the admin section.
