+++ src/features/admin/withdrawal/index.tsx
@@ -13,6 +13,7 @@ import { WithdrawalStatus } from '@/hooks/api/withdrawal/enums'
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@/components/ui/select'
 import type { WithdrawalFilter } from '@/hooks/api/withdrawal/types'
 import { formatDateToString } from '@/helpers/datetime'
+import { Separator } from '@/components/ui/separator'
 
 const route = getRouteApi('/_authenticated/admin/withdrawal/')
 
@@ -23,7 +24,7 @@ export default function Withdrawal() {
     const [statusFilter, setStatusFilter] = useState<string>(WithdrawalStatus.PENDING)
     const [fromDate, setFromDate] = useState<Date>(new Date())
     const [toDate, setToDate] = useState<Date>(new Date())
-    const [page, setPage] = useState(1)
+    const [page, setPage] = useState(0)
     const [size] = useState(10)
 
     const statusOptions = [
@@ -35,7 +36,7 @@ export default function Withdrawal() {
     ]
 
     useEffect(() => {
-        setPage(1)
+        setPage(0)
     }, [statusFilter, fromDate, toDate])
 
 
@@ -59,32 +60,32 @@ export default function Withdrawal() {
         )
     }
 
-    const withdrawalRequests = data?.content || []
+    const withdrawalRequests = data?.data.data || []
 
     return (
         <WithdrawalProvider>
             <>
                 <Main>
-                    <div className='mb-2 flex flex-wrap items-center justify-between space-y-2'>
-                        <div>
+                    <div className='flex flex-col space-y-6'>
+                        {/* Header Section */}
+                        <div className='flex flex-col space-y-2'>
                             <h2 className='text-2xl font-bold tracking-tight'>Quản lý yêu cầu rút tiền</h2>
                             <p className='text-muted-foreground'>
                                 Xem và xử lý các yêu cầu rút tiền từ người dùng.
                             </p>
                         </div>
-                    </div>
 
-                    {/* Filter Section */}
-                    <div className='mb-4 rounded-md border bg-card p-4'>
-                        <div className='grid grid-cols-1 gap-4 md:grid-cols-3'>
-                            {/* Status Filter */}
+                        <Separator />
+
+                        {/* Filter Section */}
+                        <div className='grid gap-6 md:grid-cols-3 lg:grid-cols-4'>
                             <div className='space-y-2'>
-                                <Label htmlFor='status-filter'>Trạng thái</Label>
+                                <Label htmlFor='status-filter' className='text-sm font-medium'>Trạng thái</Label>
                                 <Select
                                     value={statusFilter}
                                     onValueChange={setStatusFilter}
                                 >
-                                    <SelectTrigger>
+                                    <SelectTrigger className='w-full bg-background'>
                                         <SelectValue placeholder='Chọn trạng thái' />
                                     </SelectTrigger>
                                     <SelectContent>
@@ -97,38 +98,40 @@ export default function Withdrawal() {
                                 </Select>
                             </div>
 
-                            {/* From Date Filter */}
                             <div className='space-y-2'>
-                                <Label htmlFor='from-date'>Từ ngày</Label>
-                                <DatePicker
-                                    selected={fromDate}
-                                    onSelect={(date) => setFromDate(date || new Date())}
-                                    placeholder='Chọn ngày bắt đầu'
-                                    minDate={undefined}
-                                />
+                                <Label className='text-sm font-medium'>Từ ngày</Label>
+                                <div className='[&>button]:w-full'>
+                                    <DatePicker
+                                        selected={fromDate}
+                                        onSelect={(date) => setFromDate(date || new Date())}
+                                        placeholder='Chọn ngày bắt đầu'
+                                    />
+                                </div>
                             </div>
 
-                            {/* To Date Filter */}
                             <div className='space-y-2'>
-                                <Label htmlFor='to-date'>Đến ngày</Label>
-                                <DatePicker
-                                    selected={toDate}
-                                    onSelect={(date) => setToDate(date || new Date())}
-                                    placeholder='Chọn ngày kết thúc'
-                                    minDate={fromDate}
-                                />
+                                <Label className='text-sm font-medium'>Đến ngày</Label>
+                                <div className='[&>button]:w-full'>
+                                    <DatePicker
+                                        selected={toDate}
+                                        onSelect={(date) => setToDate(date || new Date())}
+                                        placeholder='Chọn ngày kết thúc'
+                                        minDate={fromDate}
+                                    />
+                                </div>
                             </div>
                         </div>
-                    </div>
 
-                    <div className='-mx-4 flex-1 overflow-auto px-4 py-1 lg:flex-row lg:space-y-0 lg:space-x-12'>
-                        {isLoading ? (
-                            <div className='flex items-center justify-center py-12'>
-                                <div className='text-muted-foreground'>Đang tải dữ liệu...</div>
-                            </div>
-                        ) : (
-                            <WithdrawalTable data={withdrawalRequests} search={search} navigate={navigate} />
-                        )}
+                        {/* Table Section */}
+                        <div className='flex-1'>
+                            {isLoading ? (
+                                <div className='flex h-64 items-center justify-center rounded-md border border-dashed bg-muted/20'>
+                                    <div className='text-muted-foreground'>Đang tải dữ liệu...</div>
+                                </div>
+                            ) : (
+                                <WithdrawalTable data={withdrawalRequests} search={search} navigate={navigate} />
+                            )}
+                        </div>
                     </div>
                 </Main>
 
@@ -137,4 +140,3 @@ export default function Withdrawal() {
         </WithdrawalProvider>
     )
 }
-
