+++ src/routeTree.gen.ts
@@ -72,15 +72,12 @@ import { Route as AuthenticatedCompanyManageProjectsIndexRouteImport } from './r
 import { Route as AuthenticatedCompanyManageMyTransactionsIndexRouteImport } from './routes/_authenticated/company-manage/my-transactions/index'
 import { Route as AuthenticatedAdminWithdrawalIndexRouteImport } from './routes/_authenticated/admin/withdrawal/index'
 import { Route as AuthenticatedAdminUsersIndexRouteImport } from './routes/_authenticated/admin/users/index'
-import { Route as AuthenticatedAdminTemplatesIndexRouteImport } from './routes/_authenticated/admin/templates/index'
-import { Route as AuthenticatedAdminTasksIndexRouteImport } from './routes/_authenticated/admin/tasks/index'
 import { Route as AuthenticatedAdminSystemConfigIndexRouteImport } from './routes/_authenticated/admin/system-config/index'
 import { Route as AuthenticatedAdminSkillsIndexRouteImport } from './routes/_authenticated/admin/skills/index'
 import { Route as AuthenticatedAdminSettingsIndexRouteImport } from './routes/_authenticated/admin/settings/index'
 import { Route as AuthenticatedAdminReportsIndexRouteImport } from './routes/_authenticated/admin/reports/index'
 import { Route as AuthenticatedAdminProjectsIndexRouteImport } from './routes/_authenticated/admin/projects/index'
 import { Route as AuthenticatedAdminCompanyChecklistConfigIndexRouteImport } from './routes/_authenticated/admin/company-checklist-config/index'
-import { Route as AuthenticatedAdminCompaniesIndexRouteImport } from './routes/_authenticated/admin/companies/index'
 import { Route as AuthenticatedTalentSettingsAppearanceRouteImport } from './routes/_authenticated/talent/settings/appearance'
 import { Route as AuthenticatedTalentSettingsAccountRouteImport } from './routes/_authenticated/talent/settings/account'
 import { Route as AuthenticatedMentorSettingsAppearanceRouteImport } from './routes/_authenticated/mentor/settings/appearance'
@@ -94,6 +91,7 @@ import { Route as AuthenticatedAdminSettingsAccountRouteImport } from './routes/
 import { Route as AuthenticatedTalentProjectsProjectIdIndexRouteImport } from './routes/_authenticated/talent/projects/$projectId/index'
 import { Route as AuthenticatedMentorProjectsProjectIdIndexRouteImport } from './routes/_authenticated/mentor/projects/$projectId/index'
 import { Route as AuthenticatedLabAdminUsersUserIdIndexRouteImport } from './routes/_authenticated/lab-admin/users/$userId/index'
+import { Route as AuthenticatedLabAdminRequestsRequestIdIndexRouteImport } from './routes/_authenticated/lab-admin/requests/$requestId/index'
 import { Route as AuthenticatedLabAdminProjectsProjectIdIndexRouteImport } from './routes/_authenticated/lab-admin/projects/$projectId/index'
 import { Route as AuthenticatedLabAdminCompaniesCompanyIdIndexRouteImport } from './routes/_authenticated/lab-admin/companies/$companyId/index'
 import { Route as AuthenticatedCompanyManageProjectsCreateIndexRouteImport } from './routes/_authenticated/company-manage/projects/create/index'
@@ -101,7 +99,6 @@ import { Route as AuthenticatedCompanyManageProjectsProjectIdIndexRouteImport } 
 import { Route as AuthenticatedAdminUsersCreateIndexRouteImport } from './routes/_authenticated/admin/users/create/index'
 import { Route as AuthenticatedAdminUsersUserIdIndexRouteImport } from './routes/_authenticated/admin/users/$userId/index'
 import { Route as AuthenticatedAdminProjectsProjectIdIndexRouteImport } from './routes/_authenticated/admin/projects/$projectId/index'
-import { Route as AuthenticatedAdminCompaniesCompanyIdIndexRouteImport } from './routes/_authenticated/admin/companies/$companyId/index'
 import { Route as AuthenticatedTalentProjectsProjectIdTemplatesIndexRouteImport } from './routes/_authenticated/talent/projects/$projectId/templates/index'
 import { Route as AuthenticatedTalentProjectsProjectIdMembersIndexRouteImport } from './routes/_authenticated/talent/projects/$projectId/members/index'
 import { Route as AuthenticatedTalentProjectsProjectIdFinancialsIndexRouteImport } from './routes/_authenticated/talent/projects/$projectId/financials/index'
@@ -485,18 +482,6 @@ const AuthenticatedAdminUsersIndexRoute =
     path: '/users/',
     getParentRoute: () => AuthenticatedAdminRouteRoute,
   } as any)
-const AuthenticatedAdminTemplatesIndexRoute =
-  AuthenticatedAdminTemplatesIndexRouteImport.update({
-    id: '/templates/',
-    path: '/templates/',
-    getParentRoute: () => AuthenticatedAdminRouteRoute,
-  } as any)
-const AuthenticatedAdminTasksIndexRoute =
-  AuthenticatedAdminTasksIndexRouteImport.update({
-    id: '/tasks/',
-    path: '/tasks/',
-    getParentRoute: () => AuthenticatedAdminRouteRoute,
-  } as any)
 const AuthenticatedAdminSystemConfigIndexRoute =
   AuthenticatedAdminSystemConfigIndexRouteImport.update({
     id: '/system-config/',
@@ -531,12 +516,6 @@ const AuthenticatedAdminCompanyChecklistConfigIndexRoute =
   AuthenticatedAdminCompanyChecklistConfigIndexRouteImport.update({
     id: '/company-checklist-config/',
     path: '/company-checklist-config/',
-    getParentRoute: () => AuthenticatedAdminRouteRoute,
-  } as any)
-const AuthenticatedAdminCompaniesIndexRoute =
-  AuthenticatedAdminCompaniesIndexRouteImport.update({
-    id: '/companies/',
-    path: '/companies/',
     getParentRoute: () => AuthenticatedAdminRouteRoute,
   } as any)
 const AuthenticatedTalentSettingsAppearanceRoute =
@@ -617,6 +596,12 @@ const AuthenticatedLabAdminUsersUserIdIndexRoute =
     path: '/users/$userId/',
     getParentRoute: () => AuthenticatedLabAdminRouteRoute,
   } as any)
+const AuthenticatedLabAdminRequestsRequestIdIndexRoute =
+  AuthenticatedLabAdminRequestsRequestIdIndexRouteImport.update({
+    id: '/requests/$requestId/',
+    path: '/requests/$requestId/',
+    getParentRoute: () => AuthenticatedLabAdminRouteRoute,
+  } as any)
 const AuthenticatedLabAdminProjectsProjectIdIndexRoute =
   AuthenticatedLabAdminProjectsProjectIdIndexRouteImport.update({
     id: '/projects/$projectId/',
@@ -657,12 +642,6 @@ const AuthenticatedAdminProjectsProjectIdIndexRoute =
   AuthenticatedAdminProjectsProjectIdIndexRouteImport.update({
     id: '/projects/$projectId/',
     path: '/projects/$projectId/',
-    getParentRoute: () => AuthenticatedAdminRouteRoute,
-  } as any)
-const AuthenticatedAdminCompaniesCompanyIdIndexRoute =
-  AuthenticatedAdminCompaniesCompanyIdIndexRouteImport.update({
-    id: '/companies/$companyId/',
-    path: '/companies/$companyId/',
     getParentRoute: () => AuthenticatedAdminRouteRoute,
   } as any)
 const AuthenticatedTalentProjectsProjectIdTemplatesIndexRoute =
@@ -884,15 +863,12 @@ export interface FileRoutesByFullPath {
   '/mentor/settings/appearance': typeof AuthenticatedMentorSettingsAppearanceRoute
   '/talent/settings/account': typeof AuthenticatedTalentSettingsAccountRoute
   '/talent/settings/appearance': typeof AuthenticatedTalentSettingsAppearanceRoute
-  '/admin/companies': typeof AuthenticatedAdminCompaniesIndexRoute
   '/admin/company-checklist-config': typeof AuthenticatedAdminCompanyChecklistConfigIndexRoute
   '/admin/projects': typeof AuthenticatedAdminProjectsIndexRoute
   '/admin/reports': typeof AuthenticatedAdminReportsIndexRoute
   '/admin/settings/': typeof AuthenticatedAdminSettingsIndexRoute
   '/admin/skills': typeof AuthenticatedAdminSkillsIndexRoute
   '/admin/system-config': typeof AuthenticatedAdminSystemConfigIndexRoute
-  '/admin/tasks': typeof AuthenticatedAdminTasksIndexRoute
-  '/admin/templates': typeof AuthenticatedAdminTemplatesIndexRoute
   '/admin/users': typeof AuthenticatedAdminUsersIndexRoute
   '/admin/withdrawal': typeof AuthenticatedAdminWithdrawalIndexRoute
   '/company-manage/my-transactions': typeof AuthenticatedCompanyManageMyTransactionsIndexRoute
@@ -918,7 +894,6 @@ export interface FileRoutesByFullPath {
   '/talent/settings/': typeof AuthenticatedTalentSettingsIndexRoute
   '/talent/team-fund-distribution': typeof AuthenticatedTalentTeamFundDistributionIndexRoute
   '/talent/wallet': typeof AuthenticatedTalentWalletIndexRoute
-  '/admin/companies/$companyId': typeof AuthenticatedAdminCompaniesCompanyIdIndexRoute
   '/admin/projects/$projectId': typeof AuthenticatedAdminProjectsProjectIdIndexRoute
   '/admin/users/$userId': typeof AuthenticatedAdminUsersUserIdIndexRoute
   '/admin/users/create': typeof AuthenticatedAdminUsersCreateIndexRoute
@@ -926,6 +901,7 @@ export interface FileRoutesByFullPath {
   '/company-manage/projects/create': typeof AuthenticatedCompanyManageProjectsCreateIndexRoute
   '/lab-admin/companies/$companyId': typeof AuthenticatedLabAdminCompaniesCompanyIdIndexRoute
   '/lab-admin/projects/$projectId': typeof AuthenticatedLabAdminProjectsProjectIdIndexRoute
+  '/lab-admin/requests/$requestId': typeof AuthenticatedLabAdminRequestsRequestIdIndexRoute
   '/lab-admin/users/$userId': typeof AuthenticatedLabAdminUsersUserIdIndexRoute
   '/mentor/projects/$projectId': typeof AuthenticatedMentorProjectsProjectIdIndexRoute
   '/talent/projects/$projectId': typeof AuthenticatedTalentProjectsProjectIdIndexRoute
@@ -990,15 +966,12 @@ export interface FileRoutesByTo {
   '/mentor/settings/appearance': typeof AuthenticatedMentorSettingsAppearanceRoute
   '/talent/settings/account': typeof AuthenticatedTalentSettingsAccountRoute
   '/talent/settings/appearance': typeof AuthenticatedTalentSettingsAppearanceRoute
-  '/admin/companies': typeof AuthenticatedAdminCompaniesIndexRoute
   '/admin/company-checklist-config': typeof AuthenticatedAdminCompanyChecklistConfigIndexRoute
   '/admin/projects': typeof AuthenticatedAdminProjectsIndexRoute
   '/admin/reports': typeof AuthenticatedAdminReportsIndexRoute
   '/admin/settings': typeof AuthenticatedAdminSettingsIndexRoute
   '/admin/skills': typeof AuthenticatedAdminSkillsIndexRoute
   '/admin/system-config': typeof AuthenticatedAdminSystemConfigIndexRoute
-  '/admin/tasks': typeof AuthenticatedAdminTasksIndexRoute
-  '/admin/templates': typeof AuthenticatedAdminTemplatesIndexRoute
   '/admin/users': typeof AuthenticatedAdminUsersIndexRoute
   '/admin/withdrawal': typeof AuthenticatedAdminWithdrawalIndexRoute
   '/company-manage/my-transactions': typeof AuthenticatedCompanyManageMyTransactionsIndexRoute
@@ -1024,7 +997,6 @@ export interface FileRoutesByTo {
   '/talent/settings': typeof AuthenticatedTalentSettingsIndexRoute
   '/talent/team-fund-distribution': typeof AuthenticatedTalentTeamFundDistributionIndexRoute
   '/talent/wallet': typeof AuthenticatedTalentWalletIndexRoute
-  '/admin/companies/$companyId': typeof AuthenticatedAdminCompaniesCompanyIdIndexRoute
   '/admin/projects/$projectId': typeof AuthenticatedAdminProjectsProjectIdIndexRoute
   '/admin/users/$userId': typeof AuthenticatedAdminUsersUserIdIndexRoute
   '/admin/users/create': typeof AuthenticatedAdminUsersCreateIndexRoute
@@ -1032,6 +1004,7 @@ export interface FileRoutesByTo {
   '/company-manage/projects/create': typeof AuthenticatedCompanyManageProjectsCreateIndexRoute
   '/lab-admin/companies/$companyId': typeof AuthenticatedLabAdminCompaniesCompanyIdIndexRoute
   '/lab-admin/projects/$projectId': typeof AuthenticatedLabAdminProjectsProjectIdIndexRoute
+  '/lab-admin/requests/$requestId': typeof AuthenticatedLabAdminRequestsRequestIdIndexRoute
   '/lab-admin/users/$userId': typeof AuthenticatedLabAdminUsersUserIdIndexRoute
   '/mentor/projects/$projectId': typeof AuthenticatedMentorProjectsProjectIdIndexRoute
   '/talent/projects/$projectId': typeof AuthenticatedTalentProjectsProjectIdIndexRoute
@@ -1113,15 +1086,12 @@ export interface FileRoutesById {
   '/_authenticated/mentor/settings/appearance': typeof AuthenticatedMentorSettingsAppearanceRoute
   '/_authenticated/talent/settings/account': typeof AuthenticatedTalentSettingsAccountRoute
   '/_authenticated/talent/settings/appearance': typeof AuthenticatedTalentSettingsAppearanceRoute
-  '/_authenticated/admin/companies/': typeof AuthenticatedAdminCompaniesIndexRoute
   '/_authenticated/admin/company-checklist-config/': typeof AuthenticatedAdminCompanyChecklistConfigIndexRoute
   '/_authenticated/admin/projects/': typeof AuthenticatedAdminProjectsIndexRoute
   '/_authenticated/admin/reports/': typeof AuthenticatedAdminReportsIndexRoute
   '/_authenticated/admin/settings/': typeof AuthenticatedAdminSettingsIndexRoute
   '/_authenticated/admin/skills/': typeof AuthenticatedAdminSkillsIndexRoute
   '/_authenticated/admin/system-config/': typeof AuthenticatedAdminSystemConfigIndexRoute
-  '/_authenticated/admin/tasks/': typeof AuthenticatedAdminTasksIndexRoute
-  '/_authenticated/admin/templates/': typeof AuthenticatedAdminTemplatesIndexRoute
   '/_authenticated/admin/users/': typeof AuthenticatedAdminUsersIndexRoute
   '/_authenticated/admin/withdrawal/': typeof AuthenticatedAdminWithdrawalIndexRoute
   '/_authenticated/company-manage/my-transactions/': typeof AuthenticatedCompanyManageMyTransactionsIndexRoute
@@ -1147,7 +1117,6 @@ export interface FileRoutesById {
   '/_authenticated/talent/settings/': typeof AuthenticatedTalentSettingsIndexRoute
   '/_authenticated/talent/team-fund-distribution/': typeof AuthenticatedTalentTeamFundDistributionIndexRoute
   '/_authenticated/talent/wallet/': typeof AuthenticatedTalentWalletIndexRoute
-  '/_authenticated/admin/companies/$companyId/': typeof AuthenticatedAdminCompaniesCompanyIdIndexRoute
   '/_authenticated/admin/projects/$projectId/': typeof AuthenticatedAdminProjectsProjectIdIndexRoute
   '/_authenticated/admin/users/$userId/': typeof AuthenticatedAdminUsersUserIdIndexRoute
   '/_authenticated/admin/users/create/': typeof AuthenticatedAdminUsersCreateIndexRoute
@@ -1155,6 +1124,7 @@ export interface FileRoutesById {
   '/_authenticated/company-manage/projects/create/': typeof AuthenticatedCompanyManageProjectsCreateIndexRoute
   '/_authenticated/lab-admin/companies/$companyId/': typeof AuthenticatedLabAdminCompaniesCompanyIdIndexRoute
   '/_authenticated/lab-admin/projects/$projectId/': typeof AuthenticatedLabAdminProjectsProjectIdIndexRoute
+  '/_authenticated/lab-admin/requests/$requestId/': typeof AuthenticatedLabAdminRequestsRequestIdIndexRoute
   '/_authenticated/lab-admin/users/$userId/': typeof AuthenticatedLabAdminUsersUserIdIndexRoute
   '/_authenticated/mentor/projects/$projectId/': typeof AuthenticatedMentorProjectsProjectIdIndexRoute
   '/_authenticated/talent/projects/$projectId/': typeof AuthenticatedTalentProjectsProjectIdIndexRoute
@@ -1234,15 +1204,12 @@ export interface FileRouteTypes {
     | '/mentor/settings/appearance'
     | '/talent/settings/account'
     | '/talent/settings/appearance'
-    | '/admin/companies'
     | '/admin/company-checklist-config'
     | '/admin/projects'
     | '/admin/reports'
     | '/admin/settings/'
     | '/admin/skills'
     | '/admin/system-config'
-    | '/admin/tasks'
-    | '/admin/templates'
     | '/admin/users'
     | '/admin/withdrawal'
     | '/company-manage/my-transactions'
@@ -1268,7 +1235,6 @@ export interface FileRouteTypes {
     | '/talent/settings/'
     | '/talent/team-fund-distribution'
     | '/talent/wallet'
-    | '/admin/companies/$companyId'
     | '/admin/projects/$projectId'
     | '/admin/users/$userId'
     | '/admin/users/create'
@@ -1276,6 +1242,7 @@ export interface FileRouteTypes {
     | '/company-manage/projects/create'
     | '/lab-admin/companies/$companyId'
     | '/lab-admin/projects/$projectId'
+    | '/lab-admin/requests/$requestId'
     | '/lab-admin/users/$userId'
     | '/mentor/projects/$projectId'
     | '/talent/projects/$projectId'
@@ -1340,15 +1307,12 @@ export interface FileRouteTypes {
     | '/mentor/settings/appearance'
     | '/talent/settings/account'
     | '/talent/settings/appearance'
-    | '/admin/companies'
     | '/admin/company-checklist-config'
     | '/admin/projects'
     | '/admin/reports'
     | '/admin/settings'
     | '/admin/skills'
     | '/admin/system-config'
-    | '/admin/tasks'
-    | '/admin/templates'
     | '/admin/users'
     | '/admin/withdrawal'
     | '/company-manage/my-transactions'
@@ -1374,7 +1338,6 @@ export interface FileRouteTypes {
     | '/talent/settings'
     | '/talent/team-fund-distribution'
     | '/talent/wallet'
-    | '/admin/companies/$companyId'
     | '/admin/projects/$projectId'
     | '/admin/users/$userId'
     | '/admin/users/create'
@@ -1382,6 +1345,7 @@ export interface FileRouteTypes {
     | '/company-manage/projects/create'
     | '/lab-admin/companies/$companyId'
     | '/lab-admin/projects/$projectId'
+    | '/lab-admin/requests/$requestId'
     | '/lab-admin/users/$userId'
     | '/mentor/projects/$projectId'
     | '/talent/projects/$projectId'
@@ -1462,15 +1426,12 @@ export interface FileRouteTypes {
     | '/_authenticated/mentor/settings/appearance'
     | '/_authenticated/talent/settings/account'
     | '/_authenticated/talent/settings/appearance'
-    | '/_authenticated/admin/companies/'
     | '/_authenticated/admin/company-checklist-config/'
     | '/_authenticated/admin/projects/'
     | '/_authenticated/admin/reports/'
     | '/_authenticated/admin/settings/'
     | '/_authenticated/admin/skills/'
     | '/_authenticated/admin/system-config/'
-    | '/_authenticated/admin/tasks/'
-    | '/_authenticated/admin/templates/'
     | '/_authenticated/admin/users/'
     | '/_authenticated/admin/withdrawal/'
     | '/_authenticated/company-manage/my-transactions/'
@@ -1496,7 +1457,6 @@ export interface FileRouteTypes {
     | '/_authenticated/talent/settings/'
     | '/_authenticated/talent/team-fund-distribution/'
     | '/_authenticated/talent/wallet/'
-    | '/_authenticated/admin/companies/$companyId/'
     | '/_authenticated/admin/projects/$projectId/'
     | '/_authenticated/admin/users/$userId/'
     | '/_authenticated/admin/users/create/'
@@ -1504,6 +1464,7 @@ export interface FileRouteTypes {
     | '/_authenticated/company-manage/projects/create/'
     | '/_authenticated/lab-admin/companies/$companyId/'
     | '/_authenticated/lab-admin/projects/$projectId/'
+    | '/_authenticated/lab-admin/requests/$requestId/'
     | '/_authenticated/lab-admin/users/$userId/'
     | '/_authenticated/mentor/projects/$projectId/'
     | '/_authenticated/talent/projects/$projectId/'
@@ -1995,20 +1956,6 @@ declare module '@tanstack/react-router' {
       preLoaderRoute: typeof AuthenticatedAdminUsersIndexRouteImport
       parentRoute: typeof AuthenticatedAdminRouteRoute
     }
-    '/_authenticated/admin/templates/': {
-      id: '/_authenticated/admin/templates/'
-      path: '/templates'
-      fullPath: '/admin/templates'
-      preLoaderRoute: typeof AuthenticatedAdminTemplatesIndexRouteImport
-      parentRoute: typeof AuthenticatedAdminRouteRoute
-    }
-    '/_authenticated/admin/tasks/': {
-      id: '/_authenticated/admin/tasks/'
-      path: '/tasks'
-      fullPath: '/admin/tasks'
-      preLoaderRoute: typeof AuthenticatedAdminTasksIndexRouteImport
-      parentRoute: typeof AuthenticatedAdminRouteRoute
-    }
     '/_authenticated/admin/system-config/': {
       id: '/_authenticated/admin/system-config/'
       path: '/system-config'
@@ -2049,13 +1996,6 @@ declare module '@tanstack/react-router' {
       path: '/company-checklist-config'
       fullPath: '/admin/company-checklist-config'
       preLoaderRoute: typeof AuthenticatedAdminCompanyChecklistConfigIndexRouteImport
-      parentRoute: typeof AuthenticatedAdminRouteRoute
-    }
-    '/_authenticated/admin/companies/': {
-      id: '/_authenticated/admin/companies/'
-      path: '/companies'
-      fullPath: '/admin/companies'
-      preLoaderRoute: typeof AuthenticatedAdminCompaniesIndexRouteImport
       parentRoute: typeof AuthenticatedAdminRouteRoute
     }
     '/_authenticated/talent/settings/appearance': {
@@ -2149,6 +2089,13 @@ declare module '@tanstack/react-router' {
       preLoaderRoute: typeof AuthenticatedLabAdminUsersUserIdIndexRouteImport
       parentRoute: typeof AuthenticatedLabAdminRouteRoute
     }
+    '/_authenticated/lab-admin/requests/$requestId/': {
+      id: '/_authenticated/lab-admin/requests/$requestId/'
+      path: '/requests/$requestId'
+      fullPath: '/lab-admin/requests/$requestId'
+      preLoaderRoute: typeof AuthenticatedLabAdminRequestsRequestIdIndexRouteImport
+      parentRoute: typeof AuthenticatedLabAdminRouteRoute
+    }
     '/_authenticated/lab-admin/projects/$projectId/': {
       id: '/_authenticated/lab-admin/projects/$projectId/'
       path: '/projects/$projectId'
@@ -2196,13 +2143,6 @@ declare module '@tanstack/react-router' {
       path: '/projects/$projectId'
       fullPath: '/admin/projects/$projectId'
       preLoaderRoute: typeof AuthenticatedAdminProjectsProjectIdIndexRouteImport
-      parentRoute: typeof AuthenticatedAdminRouteRoute
-    }
-    '/_authenticated/admin/companies/$companyId/': {
-      id: '/_authenticated/admin/companies/$companyId/'
-      path: '/companies/$companyId'
-      fullPath: '/admin/companies/$companyId'
-      preLoaderRoute: typeof AuthenticatedAdminCompaniesCompanyIdIndexRouteImport
       parentRoute: typeof AuthenticatedAdminRouteRoute
     }
     '/_authenticated/talent/projects/$projectId/templates/': {
@@ -2465,17 +2405,13 @@ const AuthenticatedAdminSettingsRouteRouteWithChildren =
 interface AuthenticatedAdminRouteRouteChildren {
   AuthenticatedAdminSettingsRouteRoute: typeof AuthenticatedAdminSettingsRouteRouteWithChildren
   AuthenticatedAdminIndexRoute: typeof AuthenticatedAdminIndexRoute
-  AuthenticatedAdminCompaniesIndexRoute: typeof AuthenticatedAdminCompaniesIndexRoute
   AuthenticatedAdminCompanyChecklistConfigIndexRoute: typeof AuthenticatedAdminCompanyChecklistConfigIndexRoute
   AuthenticatedAdminProjectsIndexRoute: typeof AuthenticatedAdminProjectsIndexRoute
   AuthenticatedAdminReportsIndexRoute: typeof AuthenticatedAdminReportsIndexRoute
   AuthenticatedAdminSkillsIndexRoute: typeof AuthenticatedAdminSkillsIndexRoute
   AuthenticatedAdminSystemConfigIndexRoute: typeof AuthenticatedAdminSystemConfigIndexRoute
-  AuthenticatedAdminTasksIndexRoute: typeof AuthenticatedAdminTasksIndexRoute
-  AuthenticatedAdminTemplatesIndexRoute: typeof AuthenticatedAdminTemplatesIndexRoute
   AuthenticatedAdminUsersIndexRoute: typeof AuthenticatedAdminUsersIndexRoute
   AuthenticatedAdminWithdrawalIndexRoute: typeof AuthenticatedAdminWithdrawalIndexRoute
-  AuthenticatedAdminCompaniesCompanyIdIndexRoute: typeof AuthenticatedAdminCompaniesCompanyIdIndexRoute
   AuthenticatedAdminProjectsProjectIdIndexRoute: typeof AuthenticatedAdminProjectsProjectIdIndexRoute
   AuthenticatedAdminUsersUserIdIndexRoute: typeof AuthenticatedAdminUsersUserIdIndexRoute
   AuthenticatedAdminUsersCreateIndexRoute: typeof AuthenticatedAdminUsersCreateIndexRoute
@@ -2490,8 +2426,6 @@ const AuthenticatedAdminRouteRouteChildren: AuthenticatedAdminRouteRouteChildren
     AuthenticatedAdminSettingsRouteRoute:
       AuthenticatedAdminSettingsRouteRouteWithChildren,
     AuthenticatedAdminIndexRoute: AuthenticatedAdminIndexRoute,
-    AuthenticatedAdminCompaniesIndexRoute:
-      AuthenticatedAdminCompaniesIndexRoute,
     AuthenticatedAdminCompanyChecklistConfigIndexRoute:
       AuthenticatedAdminCompanyChecklistConfigIndexRoute,
     AuthenticatedAdminProjectsIndexRoute: AuthenticatedAdminProjectsIndexRoute,
@@ -2499,14 +2433,9 @@ const AuthenticatedAdminRouteRouteChildren: AuthenticatedAdminRouteRouteChildren
     AuthenticatedAdminSkillsIndexRoute: AuthenticatedAdminSkillsIndexRoute,
     AuthenticatedAdminSystemConfigIndexRoute:
       AuthenticatedAdminSystemConfigIndexRoute,
-    AuthenticatedAdminTasksIndexRoute: AuthenticatedAdminTasksIndexRoute,
-    AuthenticatedAdminTemplatesIndexRoute:
-      AuthenticatedAdminTemplatesIndexRoute,
     AuthenticatedAdminUsersIndexRoute: AuthenticatedAdminUsersIndexRoute,
     AuthenticatedAdminWithdrawalIndexRoute:
       AuthenticatedAdminWithdrawalIndexRoute,
-    AuthenticatedAdminCompaniesCompanyIdIndexRoute:
-      AuthenticatedAdminCompaniesCompanyIdIndexRoute,
     AuthenticatedAdminProjectsProjectIdIndexRoute:
       AuthenticatedAdminProjectsProjectIdIndexRoute,
     AuthenticatedAdminUsersUserIdIndexRoute:
@@ -2632,6 +2561,7 @@ interface AuthenticatedLabAdminRouteRouteChildren {
   AuthenticatedLabAdminUsersIndexRoute: typeof AuthenticatedLabAdminUsersIndexRoute
   AuthenticatedLabAdminCompaniesCompanyIdIndexRoute: typeof AuthenticatedLabAdminCompaniesCompanyIdIndexRoute
   AuthenticatedLabAdminProjectsProjectIdIndexRoute: typeof AuthenticatedLabAdminProjectsProjectIdIndexRoute
+  AuthenticatedLabAdminRequestsRequestIdIndexRoute: typeof AuthenticatedLabAdminRequestsRequestIdIndexRoute
   AuthenticatedLabAdminUsersUserIdIndexRoute: typeof AuthenticatedLabAdminUsersUserIdIndexRoute
   AuthenticatedLabAdminProjectsProjectIdMilestoneIdIndexRoute: typeof AuthenticatedLabAdminProjectsProjectIdMilestoneIdIndexRoute
   AuthenticatedLabAdminProjectsProjectIdMembersIndexRoute: typeof AuthenticatedLabAdminProjectsProjectIdMembersIndexRoute
@@ -2661,6 +2591,8 @@ const AuthenticatedLabAdminRouteRouteChildren: AuthenticatedLabAdminRouteRouteCh
       AuthenticatedLabAdminCompaniesCompanyIdIndexRoute,
     AuthenticatedLabAdminProjectsProjectIdIndexRoute:
       AuthenticatedLabAdminProjectsProjectIdIndexRoute,
+    AuthenticatedLabAdminRequestsRequestIdIndexRoute:
+      AuthenticatedLabAdminRequestsRequestIdIndexRoute,
     AuthenticatedLabAdminUsersUserIdIndexRoute:
       AuthenticatedLabAdminUsersUserIdIndexRoute,
     AuthenticatedLabAdminProjectsProjectIdMilestoneIdIndexRoute:
