--- src/routes/_authenticated/admin/tasks/index.tsx
@@ -1,7 +1,0 @@
-import Task from '@/features/admin/tasks'
-import { createFileRoute } from '@tanstack/react-router'
-
-export const Route = createFileRoute('/_authenticated/admin/tasks/')({
-  component: Task,
-})
-
