--- src/routes/_authenticated/admin/templates/index.tsx
@@ -1,6 +1,0 @@
-import { createFileRoute } from '@tanstack/react-router'
-import TemplatesManagementPage from '@/features/labAdmin/templates'
-
-export const Route = createFileRoute('/_authenticated/admin/templates/')({
-    component: TemplatesManagementPage,
-})
